requirements.txt
streamlit
opencv-python
ultralytics
numpy
Pillow
google-genai
streamlit-webrtc
av